% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction.R
\name{download_file}
\alias{download_file}
\title{Download file from \href{https://www.reachresourcecentre.info/}{REACH Resource Centre (RC)}}
\usage{
download_file(file_url, destination_folder)
}
\arguments{
\item{file_url}{string URL to file}

\item{destination_folder}{string folder path to the destination folder where the data will be downloaded. If the folder does not exist, it will be created}
}
\value{
the file from the file_url
}
\description{
Download file from \href{https://www.reachresourcecentre.info/}{REACH Resource Centre (RC)}
}
\examples{
\dontrun{
file_url <- RC_extract_onePage("https://www.reachresourcecentre.info/search/?search=1&initiative\%5B\%5D=reach&ptheme\%5B\%5D=cash&dates=&keywords=JMMI")[1]
download_file(file_url, "data")
}
}
