% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction.R
\name{download_files}
\alias{download_files}
\title{Download files from list of URLs from \href{https://www.reachresourcecentre.info/}{REACH Resource Centre (RC)}}
\usage{
download_files(url_list, destination_folder)
}
\arguments{
\item{url_list}{list of URLs to files from the RC}

\item{destination_folder}{string folder path to the destination folder where the data will be downloaded. If the folder does not exist, it will be created}
}
\value{
files from the file_url
}
\description{
Download files from list of URLs from \href{https://www.reachresourcecentre.info/}{REACH Resource Centre (RC)}
}
\examples{
\dontrun{
file_urls <- RC_extract_onePage("https://www.reachresourcecentre.info/search/?search=1&initiative\%5B\%5D=reach&ptheme\%5B\%5D=cash&dates=&keywords=JMMI")
download_files(file_urls, "data")
}
}
