% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction.R
\name{RC_extract_allPages}
\alias{RC_extract_allPages}
\title{Extract download links from all pages from a search URL in \href{https://www.reachresourcecentre.info/}{REACH Resource Centre (RC)}}
\usage{
RC_extract_allPages(search_URL)
}
\arguments{
\item{search_URL}{string with a search URL from REACH RC.}
}
\value{
list of links to files from the search
}
\description{
Extract download links from all pages from a search URL in \href{https://www.reachresourcecentre.info/}{REACH Resource Centre (RC)}
}
\examples{
RC_extract_allPages("https://www.reachresourcecentre.info/search/?search=1&initiative\%5B\%5D=reach&ptheme\%5B\%5D=cash&dates=&keywords=JMMI")

}
